Put all the code in a main() function
It's a lot easier to understand your code if you put things in functions. Conventionally, a Python script has a `main` function that does everything, and it's invoked with

        if __name__ == '__main__':
            main()

So at a minimum, I'd suggest wrapping all your code (except for the `solver` function) in a `main` function, as done in this commit.