Split the for loop out into a separate function
Continuing from the previous commit, which made the main() function, it's also good practice to break each logical piece of calculation into its own function.
Specifically, the for loop in the middle.
You can then give it a name that clearly identifies what it does, like find_solutions.

This new function will need access to the variables middlenumber and successes, so you have to take them out of the beginning of main() and put them in the new find_solutions() function.
That's a good thing, since variables should be declared right before they get used.
It helps minimize the number of things you need to keep in your head at once.