Improve the naming in the (former) solver() function
solver is not a great name.
It would be a decent name if it actually solved anything.
But it doesn't.
That function actually checks whether two numbers and their difference contain all the digits from 1 to 9 a specified number of times.
So give it a name that reflects that.
You can expand on the name by using a doc comment.

Similarly, number and number2 aren't very conventional names, although in this case because you're just running a mathematical algorithm, there aren't much better choices.
It's common to use a and b for things like this, since the numbers are integers, or x and y if they were floats.
And it probably makes more sense to pass the number of times each digit should appear as the argument, not the number of digits in each of the original numbers, because that way it's a lot easier to understand what the code does.