Return None instead of False to indicate a failed check
It's conventional to use None as the return value to indicate that a function doesn't have any meaningful result to return, so I'd suggest using that instead of False.
Nothing wrong with False, but it's a bit strange to have a function returning either a boolean or a string depending on the result of a calculation.