Combine if statements in main()
You can combine the if statements testing the number of digits in your main() function.
It's also better to use elif to show that the options are mutually exclusive.
Then you can finish it up with an else block to catch the case where the user enters a number you can't handle.