Just fixing a few spacing issues
Python has a standard, PEP 8, for how code should be spaced.
It's not required, but it does tend to make code look pretty good and easy to read.
Since readability of the source code in Python is important (it works as documentation), it's good to get used to the conventions.