Remove output from find_solutions()
It's a good idea in most cases to separate the calculating parts of your code (sometimes called "business logic") from the input and output.
So you can the print statements from the find_solutions() function, which does most of the calculating.
This does mean you lose the printout of the individual solutions.
Don't worry, we'll get it right back.