Move computation of mins and maxes into find_solutions()
Speaking of separating calculation from input/output, the calculation of the mins and maxes is really part of the process of finding solutions.
It belongs with the loop, in find_solutions(), not in main().
Making that change means you get to pass find_solutions one argument, the number of digits, not 4, which is probably a good sign.