Iterate over permutations instead of all n-digit numbers
One way to avoid taking a lot of time checking whether all digits are used the right number of times is to start with the right set of digits and iterate over all permutations of it.
Python has a built-in tool for that, the permutations() function of the itertools module.
It takes a sequence like '123456789' and goes through all possible orderings of the elements in the sequence.
All you have to do for each ordering is covert the first three digits into a number, the second three digits into a number, and the third three digits into a number, and see if the first number minus the second equals the third.
To avoid giving duplicates, you can only test the case where the three numbers are produced in decreasing order.

This slightly improves on the initial results.
The profiler gives this output:

         1451702 function calls in 0.999 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.999    0.999 <string>:1(<module>)
        1    0.000    0.000    0.999    0.999 impossibleproblem.py:16(main)
        1    0.626    0.626    0.999    0.999 impossibleproblem.py:4(find_solutions)
        1    0.000    0.000    0.999    0.999 {built-in method exec}
   362880    0.026    0.000    0.026    0.000 {built-in method iter}
        3    0.000    0.000    0.000    0.000 {built-in method len}
        4    0.000    0.000    0.000    0.000 {built-in method print}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
      168    0.000    0.000    0.000    0.000 {method 'extend' of 'list' objects}
        2    0.000    0.000    0.000    0.000 {method 'format' of 'str' objects}
  1088640    0.347    0.000    0.347    0.000 {method 'join' of 'str' objects}