Use a 3-element frozenset to represent each solution
One confusing thing about your implementation, by the way, is that it adds two numbers to the list of "solutions" for each solution that it actually finds.
Why those two, and not the third one?
More importantly, why should each solution count as two separate elements to the list?
Since you're looking for triplets of numbers which satisfy certain criteria, you should represent each solution by a triple, and each triple should be only one element in the final list.
The builtin frozenset class is perfect for that; it's a set (so it doesn't care about the order of its elements) which can't be modified.
Making this change shows that you actually have only 168 solutions; that is, 168 ways of splitting 9 digits between three 3-digit numbers such that one is the difference between the other two.
You were basically double-counting solutions before, since you added two things to the list for each solution.

This isn't a timing issue, but it doesn't hurt to check the profiler and make sure your code doesn't get slower.
Here is the output:

         1451702 function calls in 1.019 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    1.019    1.019 <string>:1(<module>)
        1    0.000    0.000    1.019    1.019 impossibleproblem.py:16(main)
        1    0.514    0.514    1.019    1.019 impossibleproblem.py:4(find_solutions)
        1    0.000    0.000    1.019    1.019 {built-in method exec}
   362880    0.027    0.000    0.027    0.000 {built-in method iter}
        3    0.000    0.000    0.000    0.000 {built-in method len}
        4    0.000    0.000    0.000    0.000 {built-in method print}
      168    0.000    0.000    0.000    0.000 {method 'append' of 'list' objects}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
        2    0.000    0.000    0.000    0.000 {method 'format' of 'str' objects}
  1088640    0.477    0.000    0.477    0.000 {method 'join' of 'str' objects}