Use integer manipulation instead of string manipulation
Looking at the timing results of both the permutation algorithm and the counter algorithm, the main lesson to take away is that string manipulation is slow.
If you can replace that with arithmetic, you can potentially make the calculation a lot faster.
That's exactly what vaeta's digit manipulation algorithm does.
It represents each digit in the calculation as an integer and allows Python to determine which combinations are allowed using addition.
This makes the program fully 20 times faster than the counter version, and 40 times faster than the permutation version.
Here are the timing results:

         896 function calls in 0.025 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.025    0.025 <string>:1(<module>)
      356    0.023    0.000    0.023    0.000 impossibleproblem.py:11(<listcomp>)
        1    0.000    0.000    0.025    0.025 impossibleproblem.py:17(main)
      169    0.000    0.000    0.025    0.000 impossibleproblem.py:3(digits_solver)
        1    0.000    0.000    0.000    0.000 impossibleproblem.py:4(<listcomp>)
        1    0.000    0.000    0.000    0.000 impossibleproblem.py:7(<listcomp>)
       28    0.001    0.000    0.001    0.000 impossibleproblem.py:9(<listcomp>)
        1    0.000    0.000    0.025    0.025 {built-in method exec}
      163    0.000    0.000    0.000    0.000 {built-in method len}
        4    0.000    0.000    0.000    0.000 {built-in method print}
      168    0.000    0.000    0.000    0.000 {method 'add' of 'set' objects}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
        2    0.000    0.000    0.000    0.000 {method 'format' of 'str' objects}